modified the delegate methods

+++ Sources/Castled/Inbox/Model/CastledInboxItem.swift
@@ -20,7 +20,7 @@ import UIKit
     public var isRead: Bool
     public var isPinned: Bool
     public var message: [String: Any]
-    public var messageId: Int
+    public var messageId: Int64
     public var startTs: Int64
     public var sourceContext: String
     public var tag: String
@@ -48,7 +48,7 @@ import UIKit
     public required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         self.teamID = try container.decode(Int.self, forKey: .teamID)
-        self.messageId = try container.decode(Int.self, forKey: .messageId)
+        self.messageId = try container.decode(Int64.self, forKey: .messageId)
         self.sourceContext = try container.decode(String.self, forKey: .sourceContext)
         self.startTs = try container.decodeIfPresent(Int64.self, forKey: .startTs) ?? 0
         self.updatedTime = try container.decodeIfPresent(Int64.self, forKey: .updatedTime) ?? 0
